Move stub_consensus.h to kv/test
