+++ src/kv/test/kv_serialisation.cpp
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#include "consensus/test/stub_consensus.h"
 #include "ds/logger.h"
-#include "enclave/app_interface.h"
 #include "kv/kv.h"
 #include "kv/kv_serialiser.h"
-#include "node/encryptor.h"
+#include "kv/test/stub_consensus.h"
 
 #include <doctest/doctest.h>
 #include <msgpack/msgpack.hpp>
