+++ src/kv/test/stub_consensus.h
