+++ src/node/test/history_bench.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#define PICOBENCH_IMPLEMENT
-#include "consensus/test/stub_consensus.h"
+#include "kv/test/stub_consensus.h"
 #include "node/history.h"
 
 #include <cstdlib>
 #include <ctime>
+#define PICOBENCH_IMPLEMENT
 #include <picobench/picobench.hpp>
 
 extern "C"
